+++ packages/core/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./engine/runSimulation";
+export * from "./experiments/runCounterfactualExperiment";
 export * from "./random/seededRng";
 export * from "./schema/scenario";
 export * from "./accounting/invariants";